infra for level objects, new art and demo level

+++ inc/pathfind.h
@@ -38,13 +38,7 @@ enum PathfindingResult {
 } ENUM_PACK;
 
 
-bool PATH_find( const u8 pfw, const u8 pfh, // playfield dims
-                const u8 sx, const u8 sy, // start position
-                u8 ex, const u8 ey, // goal position
-                const u8 * const attrs, // playfield attribute addr
-                const u8 stride_x, const u8 stride_y, // how to address playfield attributes (row-major or col-major)
-                const u8 wall_bit // bits that mean position is blocked
-                );
+bool PATH_find(const u8 pfw, const u8 pfh, const u8 sx, const u8 sy, u8 ex, const u8 ey, const u8*const attrs, const u16 stride_x, const u16 stride_y, const u8 wall_bit);
 
 
 PathfindingNode* PATH_node_at(u8 x, u8 y);
